fix percent

+++ cosmos/web/filters.py
@@ -54,7 +54,7 @@ def add_filters(bprint_or_app, type_='bprint'):
         a = session.query(f(getattr(Task, attribute))).join(Stage).filter(Stage.id == stage.id).scalar()
         if a is None:
             return ''
-        a = int(a)
+        a = float(a)
         if 'rss' in attribute:
             return format_memory_kb(a)
         if 'mem_req' in attribute:
